requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
openai>=1.6.1
PyPDF2==3.0.1
python-docx==1.1.0
ruff==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1